Custom costmap converter (#35)
* New custom plugin: costmap_converter_custom
Why this custom costmap_converter ?
In fact the oscillations of the robot near edges weren't due to a teb 
problem. The problem is from the costmap converter 
CostmapToPolygonsDBSMCCH (or all the others costmap converters of the 
package), in a edge, instead of considering 2 obstacle lines, it 
considers a triangle ! So there is an imaginary lethal obstacle that 
prevent the robot to cross it, creating "oscillations".
The goal of this custom costmap converter is to get rid of theses 
triangle. To to it, i'll just decompose the map into a rectangle and 3 
lines and set them as lethal obstacles.

* feasibility test

* Description of the idea

* Getting a vector of vector of float from a yaml

* setting obstacle from the yaml array

* Using the custom costmap_converter

* Simplified absolute path of map

* Garbage to be cleaned up, cleaning of the plugin incoming

* small adjustement of the map

* Deleting unsuded garbage from the plugin

* New strategy over polygons parsing for obstacles inside the map
Must work direclty on the costmap_converter package for better results, 
incoming...

* little tweak of teb_local_planner

* Finishing up cleaning the plugin

* Revert to the old strategy:
The problem of the thickness comes from teb_local_planner that i'll 
modify a bit.

* Old strategy but new static parameters

* Get rid of oscillations

* Map check offset of reso

* CI foxy-devel branch + 2cm thickness

* CostmapToPolygonsDBSMCCH in now CostmapToLines

* CI cache key update